requirements.txt
appdirs==1.4.3
CacheControl==0.12.6
certifi==2019.11.28
chardet==3.0.4
click==7.1.2
cmake==3.18.0
colorama==0.4.3
contextlib2==0.6.0
distlib==0.3.0
distro==1.4.0
dlib==19.20.0
face-recognition==1.3.0
face-recognition-models==0.3.0
h5py==2.10.0
html5lib==1.0.1
idna==2.8
imutils==0.5.3
ipaddr==2.2.0
lockfile==0.12.2
msgpack==0.6.2
numpy==1.19.1
opencv-python==4.3.0.36
packaging==20.3
pep517==0.8.2
Pillow==7.2.0
progress==1.5
pyparsing==2.4.6
pytoml==0.1.21
requests==2.22.0
retrying==1.3.3
six==1.14.0
urllib3==1.25.8
webencodings==0.5.1

scikit-learn~=0.23.2